// 11PacksIonic/src/app/model/paytm.ts
export class TxnRequest {
    MID:String;
    ORDER_ID: number;
    CUST_ID: String;
    INDUSTRY_TYPE_ID : String;
    CHANNEL_ID : String;
    TXN_AMOUNT: String;
    WEBSITE: String;
    CHECKSUMHASH:String;
    MOBILE_NO: String;
    EMAIL: String;
    CALLBACK_URL : String;
    PAYMENT_MODE_ONLY : String;
    AUTH_MODE: String;
    PAYMENT_TYPE_ID: String;
    BANK_CODE: String;
    ENVIRONMENT: String;
  }